(separate) core graphql types and queries from ee
